requirements.txt
azure-core
azure-servicebus